Various no-ops:
Add a __unused.

Make the 2byte decoder functions return 16 bits for the benefits
of picky lints.

No need to grab giant around a tsleep() when we have a timeout.

Sponsored by:	DARPA & NAI Labs.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@104060 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/geom/geom.h
@@ -295,9 +295,9 @@ extern struct sx topology_lock;
 #define GEOMGETCONF _IOWR('G',  0, struct sbuf)
 
 /* geom_enc.c */
-uint32_t g_dec_be2(u_char *p);
+uint16_t g_dec_be2(u_char *p);
 uint32_t g_dec_be4(u_char *p);
-uint32_t g_dec_le2(u_char *p);
+uint16_t g_dec_le2(u_char *p);
 uint32_t g_dec_le4(u_char *p);
 uint64_t g_dec_le8(u_char *p);
 void g_enc_le4(u_char *p, uint32_t u);
